[PATCH] mimgw: Remove Compiler Warnings
Remove some complier warnings from msvc in compat/mingw.c for value truncation from 64 bit to 32 bit integers.

Use size_t instead of int as all of the changed variables hold the result of strlen() or wcslen() which cannot be negative.
and set the size of ssize_t to 64 bit on windwos 64 bit.

Signed-off-by: Sören Krecker <soekkle@freenet.de>